backend/requirements.txt
Django>=5.1,<6.0
pandas>=2.2,<3.0
mecab-python3
psycopg2>=2.9
djangorestframework>=3.15.1
jpholiday>=0.1.10
django-cors-headers>=4.4.0